third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-getstatusforpolicy.html

<!DOCTYPE html>
<html>
    <head>
        <title>MediaKeys.getStatusForPolicy</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            async_test(function(test)
            {
                var hdcpVersionNone = "";
                var hdcpVersion2_0 = "hdcp-2.0";
                var mediaKeys;
                
                function getStatusForPolicy(hdcpVersion)
                {
                    var policy = new MediaKeysPolicy({minHdcpVersion: hdcpVersion});

                    mediaKeys.getStatusForPolicy(policy).then(function(MediaKeyStatus) {
                      // getStatusForPolicy() is not supported by Clear Key key
                      // system and the promise should always be rejected with
                      // NotSupportedError.
                      forceTestFailureFromPromise(test, "Promise resolved unexpectedly");
                    }, function(error) {
                      assert_equals(error.name, "NotSupportedError");
                      return Promise.resolve();
                    });
                }

                navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;
                    return getStatusForPolicy(hdcpVersionNone);
                }).then(function() {
                    return getStatusForPolicy(hdcpVersion2_0);
                }).then(function() {
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);                    
                });
            }, 'MediaKeys.getStatusForPolicy');
        </script>
    </body>
</html>
